recommend_Deposit/final_pjt_back/products/serializers.py
from rest_framework import serializers
from .models import Options, Products

class ProductsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Products
        fields = '__all__'
        
        
class OptionsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Options
        fields = '__all__'
        read_only_fields = ('product', )
        